add batching func

+++ packages/infura/infura.go
@@ -4,6 +4,6 @@ type Infura struct {
 	ApiKey string
 }
 
-func (i Infura) GetEtherBalance(address string) (int, error) {
+func (i Infura) GetEtherBalance(address string) (float64, error) {
 	return 11, nil
 }
